Clean up SpecialistDocumentSearchTest
This commit removes data that is not necessary for the tests to pass.

Since this is testing our expanding / allowed_values functionality,
it's better named UnifiedSearchExpandsAllowedValuesTest.
